requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
transformers
faiss-cpu
PyMuPDF
